airbloc: Refactor shared/adapter * Rename adapter to bind * Move package to package root
