+++ bind/contract.go
@@ -1,4 +1,4 @@
-package adapter
+package bind
 
 import (
 	"github.com/klaytn/klaytn/accounts/abi"
