+++ bind/contracts/app_registry.go
