+++ bind/contracts/controller_registry.go
