+++ bind/interface.go
@@ -1,4 +1,4 @@
-package adapter
+package bind
 
 import (
 	"context"
