+++ bind/managers/app_registry.go
