+++ bind/managers/controller_registry.go
