+++ bind/managers/mocks/mock_app_registry.go
