+++ bind/wrappers/app_registry.go
