+++ bind/wrappers/mocks/mock_app_registry.go
