+++ bind/wrappers/mocks/mock_controller_registry.go
